+++ app/src/main/java/jp/co/soramitsu/app/root/di/RootComponent.kt
@@ -9,9 +9,9 @@ import jp.co.soramitsu.app.root.presentation.main.di.MainFragmentComponent
 import jp.co.soramitsu.common.di.CommonApi
 import jp.co.soramitsu.common.di.scope.FeatureScope
 import jp.co.soramitsu.core_db.di.DbApi
-import jp.co.soramitsu.core_runtime.di.RuntimeApi
 import jp.co.soramitsu.feature_account_api.di.AccountFeatureApi
 import jp.co.soramitsu.feature_wallet_api.di.WalletFeatureApi
+import jp.co.soramitsu.runtime.di.RuntimeApi
 
 @Component(
     dependencies = [
