+++ app/src/main/java/jp/co/soramitsu/app/root/di/RootDependencies.kt
@@ -4,11 +4,11 @@ import jp.co.soramitsu.common.data.network.AppLinksProvider
 import jp.co.soramitsu.common.data.network.rpc.ConnectionManager
 import jp.co.soramitsu.common.mixin.api.NetworkStateMixin
 import jp.co.soramitsu.common.resources.ResourceManager
-import jp.co.soramitsu.core_runtime.runtime.RuntimeUpdater
 import jp.co.soramitsu.feature_account_api.domain.interfaces.AccountRepository
 import jp.co.soramitsu.feature_wallet_api.di.WalletUpdaters
 import jp.co.soramitsu.feature_wallet_api.domain.interfaces.WalletRepository
 import jp.co.soramitsu.feature_wallet_api.domain.model.BuyTokenRegistry
+import jp.co.soramitsu.runtime.RuntimeUpdater
 
 interface RootDependencies {
     fun networkStateMixin(): NetworkStateMixin
