+++ feature-wallet-impl/build.gradle
@@ -37,7 +37,7 @@ dependencies {
     implementation project(':common')
     implementation project(':feature-wallet-api')
     implementation project(':feature-account-api')
-    implementation project(':core-runtime')
+    implementation project(':runtime')
 
     implementation kotlinDep
