+++ runtime/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="jp.co.soramitsu.core_runtime">
+<manifest package="jp.co.soramitsu.runtime">
 
 </manifest>
