+++ runtime/src/main/assets/metadata/kusama
