+++ runtime/src/main/assets/types/kusama.json
