+++ runtime/src/main/java/jp/co/soramitsu/runtime/DefinitionsFetcher.kt
@@ -1,4 +1,4 @@
-package jp.co.soramitsu.core_runtime.runtime
+package jp.co.soramitsu.runtime
 
 import retrofit2.http.GET
 import retrofit2.http.Path
