+++ runtime/src/main/java/jp/co/soramitsu/runtime/RuntimeCache.kt
@@ -1,4 +1,4 @@
-package jp.co.soramitsu.core_runtime.runtime
+package jp.co.soramitsu.runtime
 
 import jp.co.soramitsu.common.interfaces.FileProvider
 import kotlinx.coroutines.Dispatchers
