+++ runtime/src/main/java/jp/co/soramitsu/runtime/RuntimePrepopulator.kt
@@ -1,4 +1,4 @@
-package jp.co.soramitsu.core_runtime.runtime
+package jp.co.soramitsu.runtime
 
 import android.content.Context
 import jp.co.soramitsu.common.data.storage.Preferences
