+++ settings.gradle
@@ -3,6 +3,6 @@ include ':feature-wallet-api'
 include ':feature-onboarding-impl'
 include ':feature-onboarding-api'
 include ':app', ':test-shared', ':common', ':feature-splash', 'core-db', 'core-api'
-include ':core-runtime'
+include ':runtime'
 include ':feature-account-api'
 include ':feature-account-impl'
